[INPLACE-285] service를 command/query로 분리 [25.08.03]
