+++ backend/inplace-application/build.gradle
@@ -18,4 +18,8 @@ dependencies {
     // Firebase 의존성
     implementation 'com.google.firebase:firebase-admin:9.1.1'
     implementation 'com.squareup.okhttp3:okhttp:4.11.0'
+
+    // OpenAI
+    implementation platform("org.springframework.ai:spring-ai-bom:1.0.0")
+    implementation 'org.springframework.ai:spring-ai-starter-model-openai'
 }
